% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia.R
\name{assess_eia}
\alias{assess_eia}
\title{Calculate scores for effective impervious area (EIA)}
\usage{
assess_eia()
}
\value{
List of 2 dataframes: summary and score.
}
\description{
Calculate scores for effective impervious area (EIA)
}
\examples{
assess_eia()
}
