% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trash.R
\name{assess_trash}
\alias{assess_trash}
\title{Calculate trash scores.}
\usage{
assess_trash(df_reach, reach_prefix_from_table, reach_prefix_from_layer)
}
\arguments{
\item{df_reach}{Reach data from Rapid Stream Assessment (RSA) dataset.}

\item{reach_prefix_from_table}{Field name prefix sourced from RSA reach
attribute table. Should be consistent with the name of the stream reach
attribute table from the RSA geodatabase.}

\item{reach_prefix_from_layer}{Field name prefix sourced from RSA stream
reach layer. Should be consistent with the name of the stream reach
polyline layer from the RSA geodatabase.}
}
\value{
List of 2 dataframes: summary and score.
}
\description{
Calculate trash scores.
}
\examples{
assess_trash(df_reach_example,
                       "StreamReachAttributes",
                       "StreamReaches")
}
