% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq.R
\name{assess_wq}
\alias{assess_wq}
\title{Calculate water quality scores}
\usage{
assess_wq(df, parameter_name, unit, max_criteria = NULL, min_criteria = NULL)
}
\arguments{
\item{df}{DOEE Ambient Water Quality Data that has been processed and formatted}

\item{parameter_name}{String to indicate target parameter: "Conductivity",
"Dissolved oxygen (DO)", "Escherichia coli", "pH", "Temperature, water",
"Turbidity"}

\item{unit}{String to indicate unit for water quality parameter}

\item{max_criteria}{DC water quality criteria as a maximum value}

\item{min_criteria}{DC water quality criteria as a minimum value}
}
\value{
List of three dataframes (results, summary, score)
}
\description{
Calculate water quality scores
}
\examples{
assess_wq(df_wq_processed,
                    parameter_name = "Temperature, water",
                    unit = "deg C",
                    max_criteria = 24)
}
