% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq_nutrients.R
\name{assess_wq_nutrients}
\alias{assess_wq_nutrients}
\title{Calculate water quality scores for nutrients}
\usage{
assess_wq_nutrients(
  df,
  parameter_name,
  piedmont_criteria,
  coastal_plain_criteria
)
}
\arguments{
\item{df}{DOEE Ambient Water Quality Data that has been processed and formatted}

\item{parameter_name}{String to indicate target parameter: "Nitrogen", "Phosphorus, Total (as P)"}

\item{piedmont_criteria}{Nutrient criteria recommendation for Piedmont Ecoregion 64 in mg/L}

\item{coastal_plain_criteria}{Nutrient criteria recommendation for Coastal Plain Ecoregion 65 in mg/L}
}
\value{
List of three dataframes: results, summary, score
}
\description{
Calculate water quality scores for nutrients
}
\examples{
assess_wq_nutrients(df_wq_processed_example, parameter_name = "Phosphorus, Total (as P)", piedmont_criteria = 0.04, coastal_plain_criteria = 0.0225)
}
