% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{calculate_grades}
\alias{calculate_grades}
\title{Calculate grade for each index and subwatershed}
\usage{
calculate_grades(all_scores)
}
\arguments{
\item{all_scores}{compiled scores for each metric and subwatershed}
}
\value{
List of three dataframes: index_grades, overall_grades, graphics
}
\description{
Calculate grade for each index and subwatershed
}
\examples{
calculate_grades(df_all_scores_example)
}
