% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.r
\name{create_day_seq}
\alias{create_day_seq}
\title{Calculate lagged correlation of SCI object}
\usage{
create_day_seq(n_years, start_date, leap_days = TRUE)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Calculate lagged correlation of SCI object
}
