% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{create_graphics}
\alias{create_graphics}
\title{Populate powerpoint template to create graphic for each subwatershed}
\usage{
create_graphics(df_graphics, output_path)
}
\arguments{
\item{df_graphics}{dataframe returned from \code{\link{calculate_grades}}
that includes filepaths to graphic pieces that will makeup final subwatershed
graphic}

\item{output_path}{filepath of output powerpoint slide deck (.pptx)}
}
\value{
powerpoint slide deck containing graphics for each subwatershed
}
\description{
Populate powerpoint template to create graphic for each subwatershed
}
