% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{cum_plot}
\alias{cum_plot}
\alias{cum_plot.sci_fit}
\title{Cumulative Plot}
\usage{
cum_plot(x, ...)

\method{cum_plot}{sci_fit}(x)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Cumulative Plot
}
