% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{dens_plot}
\alias{dens_plot}
\alias{dens_plot.sci_fit}
\title{Density calculation}
\usage{
dens_plot(x, ...)

\method{dens_plot}{sci_fit}(x)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Density calculation
}
