% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{format_wq}
\alias{format_wq}
\title{Format downloaded water quality data}
\usage{
format_wq(df)
}
\arguments{
\item{df}{DOEE Ambient Water Quality data downloaded from
\url{https://dcdoeepub.equisonline.com/PUBLIC.html}}
}
\value{
Formatted dataframe
}
\description{
Format downloaded water quality data
}
\examples{
format_wq(df_wq_example)
}
