% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{get_seed_connections}
\alias{get_seed_connections}
\title{get_seed_connections}
\usage{
get_seed_connections(mnidf, seed)
}
\arguments{
\item{mnidf}{N x 4 data frame of regions of seed regions and their
respective connectivity coordinates}

\item{seed}{a string indicating the seed for which connectivity
coordinates should be returned}
}
\description{
This function returns a K x 3 matrix of roi coordinates for a seed
region
}
\keyword{connectivity,}
\keyword{functional}
\keyword{regions}
\keyword{seed}
