% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.r
\name{gof_sci}
\alias{gof_sci}
\title{Goodness of fit for SCI}
\usage{
gof_sci(fit, p0.center.mass = FALSE, sci.limit = Inf)
}
\arguments{
\item{fit}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Goodness of fit for SCI
}
