% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{has_afni}
\alias{has_afni}
\title{has_afni}
\usage{
has_afni()
}
\description{
This function checks if afni is installed (at the standard location).
}
\examples{
has_afni()
}
\keyword{AFNI,}
\keyword{UNIX}
\keyword{neuroimaging,}
