% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{is_low}
\alias{is_low}
\title{is_low}
\usage{
is_low(sci, cutoff = 3.8)
}
\arguments{
\item{sci}{striatal connectivity index}

\item{cutoff}{striatal connectivity index cutoff (default is 3.8)}
}
\description{
This function tests if a given SCI is equal or lower than the
cutoff value (which defaults to 3.8 as in Sarpal 2016,
Am J Psychiatry).
}
\examples{
is_low(sci)
}
\keyword{connectivity,}
\keyword{correlation}
\keyword{functional}
\keyword{striatum,}
