% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.r
\name{lag_corr}
\alias{lag_corr}
\title{Calculate lagged correlation of SCI object}
\usage{
lag_corr(x, lag_list = seq(0, 365 * 2), method = "pearson")
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Calculate lagged correlation of SCI object
}
