% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{load_params}
\alias{load_params}
\title{load_params}
\usage{
load_params()
}
\description{
This function loads the default parameters needed for the
SCI calculation, including the default 91 MNI coordinates, the
normalization data frame, the weights vector, and an illustrative
set of 12 whole brain images for each seed.
}
\examples{
load_params()
}
\keyword{connectivity,}
\keyword{correlation}
\keyword{functional}
\keyword{striatum,}
