% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSCI.r
\name{mle_ci}
\alias{mle_ci}
\title{Confidence Intervals for MLE}
\usage{
mle_ci(j, fit, x_df, ci_n, ci, ci_method)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Confidence Intervals for MLE
}
