% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{mni2vox}
\alias{mni2vox}
\title{mni2vox}
\usage{
mni2vox(mni)
}
\arguments{
\item{mni}{- MNI coordinates}
}
\description{
This function transforms MNI coordinates into voxel coordinates.
}
\examples{
mni2vox(mni)
}
\keyword{MNI}
\keyword{neuroimaging,}
\keyword{space}
\keyword{space,}
\keyword{voxel}
