% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{normalize_sci}
\alias{normalize_sci}
\title{normalize_sci}
\usage{
normalize_sci(vals, normdf)
}
\arguments{
\item{vals}{individual raw sci connectivity values}

\item{normdf}{normalization data frame}
}
\description{
This function normalizes the N x 1 individual SCI connectivity values
using a N x 2 normalization data frame (which should comprise columns
'mean' and 'sd').
}
\examples{
normalize_sci(vals, normdf)
}
\keyword{connectivity}
\keyword{normalization,}
\keyword{striatal}
