% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{pars_afni_cmd}
\alias{pars_afni_cmd}
\title{parse_afni_cmd}
\usage{
pars_afni_cmd(params = " -ibox")
}
\arguments{
\item{params}{string of options passed to 3dmaskave (defaults to
-ibox)}
}
\description{
This function returns the full AFNI command 3dmaskave
}
\examples{
parse_afni_cmd(params)
}
