% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.r
\name{persist_test}
\alias{persist_test}
\title{Calculate lagged correlation of SCI object}
\usage{
persist_test(
  x,
  lag_list = seq(0, 365 * 2),
  method = "pearson",
  n_rep = 100,
  ci = 0.95
)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Calculate lagged correlation of SCI object
}
