% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.sci_gof}
\alias{plot.sci_gof}
\title{Plotting function for spi_fit class}
\usage{
\method{plot}{sci_fit}(x)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Plotting function for spi_fit class
}
