% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.sci_ts}
\alias{plot.sci_ts}
\title{Plotting function for spi_ts class}
\usage{
\method{plot}{sci_ts}(x, plot_type = "ts")
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Plotting function for spi_ts class
}
