% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_distributions.r
\name{ppe3}
\alias{ppe3}
\title{P function for a Pearson Type III}
\usage{
ppe3(q, shape, scale, location)
}
\arguments{
\item{x}{A sci_ts object.}

\item{window}{Time window for the moving average. In days or months, depending on the resolution of the x time series.}

\item{p0}{Whether to allow for zeroes. TRUE/FALSE. Useful for the SPI, where precipitation can be zero.}

\item{max_na_prop}{Maximum proportion of NA values within the time window. Number from 0-1. Default is 10 percent (0.1)}
}
\value{
x The rolling mean of the original time series. An sci_ts object.
}
\description{
P function for a Pearson Type III
}
