% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_proc.r
\name{pre_process}
\alias{pre_process}
\title{Generic read in function}
\usage{
pre_process(x)
}
\arguments{
\item{x}{A sci_ts object.}
}
\value{
x An sci_ts object.
}
\description{
To fill in
}
\seealso{
sci_ts
}
