% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_proc.r
\name{pre_process_monthly}
\alias{pre_process_monthly}
\title{Read in function for monthly time series}
\usage{
pre_process_monthly(x)
}
\arguments{
\item{x}{A sci_ts object.}
}
\value{
x An sci_ts object.
}
\description{
To fill in
}
\seealso{
sci_ts
}
