% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{print.sci_gof}
\alias{print.sci_gof}
\title{Print function for sci_gof class}
\usage{
\method{summary}{sci_fit}(x)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Print function for sci_gof class
}
