% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_wq}
\alias{process_wq}
\title{Process formatted water quality data}
\usage{
process_wq(df, start_date, end_date)
}
\arguments{
\item{df}{Formatted dataframe returned from \code{\link{format_wq}}}

\item{start_date}{Beginning of timeframe being evaluated (ex. first day of
five-year assessment period).}

\item{end_date}{End of timeframe being evaluated (ex. last day of
five-year assessment period).}
}
\value{
Processed dataframe
}
\description{
Process formatted water quality data
}
\examples{
process_wq(df_wq_formatted, "2015-07-01", "2020-06-30")
}
