% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{randstr}
\alias{randstr}
\title{randstr}
\usage{
randstr(n = 5000)
}
\arguments{
\item{n}{number of samples}
}
\description{
This function (from https://stackoverflow.com/questions/42734547/)
will produce a random string
}
