% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{read_nifti}
\alias{read_nifti}
\title{read_nifti}
\usage{
read_nifti(filename)
}
\arguments{
\item{filename}{filename of the nifti image (can be compressed)}
}
\description{
This function reads a nifti image using the fmri library in R
}
