% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accum_calc.r
\name{rollmean_calc}
\alias{rollmean_calc}
\title{Rolling mean calculator, calls preproccessing function}
\usage{
rollmean_calc(x, window, p0, max_na_prop = 0.1)
}
\arguments{
\item{x}{A sci_ts object.}

\item{window}{Time window for the moving average. In days or months, depending on the resolution of the x time series.}

\item{p0}{Whether to allow for zeroes. TRUE/FALSE. Useful for the SPI, where precipitation can be zero.}

\item{max_na_prop}{Maximum proportion of NA values within the time window. Number from 0-1. Default is 10 percent (0.1)}
}
\value{
x The rolling mean of the original time series. An sci_ts object.
}
\description{
Calculates a rolling mean for use in the SCI. Calls the pre-processing function automatically. No need to run pre-processing function. For an SCI without a moving window (sometimes useful for streamflow), use a window of 1.
}
\examples{
data("oh_precip_daily_ts")
example_ts <- sci_ts(ts = oh_precip_daily_ts, resolution = "daily")
roll_ts <- rollmean_calc(example_ts, window = 91, max_na_prop = 0.1, p0 = TRUE)
str(roll_ts)
head(roll_ts$df)
plot(roll_ts)
}
\seealso{
sci_ts
}
