% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\name{sci_fit}
\alias{sci_fit}
\title{Create sci_fit object}
\usage{
sci_fit(fit, roll_mean, nn, n_zero, monitor, distr, method, ref_years)
}
\arguments{
\item{fit}{}

\item{roll_mean}{}

\item{nn}{}

\item{n_zero}{}

\item{monitor}{}

\item{distr}{}

\item{method}{}

\item{ref_years}{}
}
\value{
An sci_fit object.
}
\description{
Wrapper to create an sci_fit object
}
