% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSCI.r
\name{sci_mle}
\alias{sci_mle}
\title{Wrapper for MLE fit}
\usage{
sci_mle(x_ts, resolution, distr, p0, warn, start.fun, start.fun.fix, ref_years)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Wrapper for MLE fit
}
