% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSCI.r
\name{sci_mle_j}
\alias{sci_mle_j}
\title{MLE fit for a single day or month j}
\usage{
sci_mle_j(j, x_ts, distr, x_res, p0, warn = TRUE, start.fun.fix = FALSE)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
MLE fit for a single day or month j
}
