% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.r
\name{sim_sci}
\alias{sim_sci}
\title{Calculate lagged correlation of SCI object}
\usage{
sim_sci(date_seq, accum_period)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Calculate lagged correlation of SCI object
}
