% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{summary.sci_fit}
\alias{summary.sci_fit}
\title{Summary function for sci_fit class}
\usage{
\method{summary}{sci_fit}(x, suppress = FALSE)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Summary function for sci_fit class
}
