% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{summary.sci_gof}
\alias{summary.sci_gof}
\title{Summary function for sci_gof class}
\usage{
\method{summary}{sci_fit}(x)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Summary function for sci_gof class
}
