% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.r
\name{theor_sci_corr}
\alias{theor_sci_corr}
\title{Fit SCI}
\usage{
theor_sci_corr(
  n_years,
  resolution,
  method = "pearson",
  n_rep = 100,
  window,
  lag_list = seq(0, 365 * 2),
  ci = 0.95
)
}
\arguments{
\item{window}{Fill in}

\item{x}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Fit SCI
}
