% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSCI.r
\name{transform_mle}
\alias{transform_mle}
\title{Transform SCI for MLE}
\usage{
transform_mle(
  fit,
  new_ts,
  p0.center.mass,
  ci = ci,
  ci_method = ci_method,
  ci_n = ci_n
)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Transform SCI for MLE
}
