% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSCI.r
\name{transformSCI}
\alias{transformSCI}
\title{Transform SCI}
\usage{
transformSCI(
  fit,
  new_ts = NA,
  p0.center.mass = FALSE,
  sci.limit = Inf,
  ci = 0.95,
  ci_method = "mvn",
  ci_n = 200
)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Transform SCI
}
