% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{weight_sci}
\alias{weight_sci}
\title{weight_sci}
\usage{
weight_sci(vals, weights)
}
\arguments{
\item{vals}{- individual sci connectivity values}

\item{weights}{- vector of weights}
}
\description{
This function weights the N x 1 individual SCI connectivity values
using a N x 1 vector of weights.
}
\examples{
weight_sci(vals, weights)
}
\keyword{connectivity,}
\keyword{striatal}
\keyword{weighting}
