% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSCI.r
\name{zero_mle}
\alias{zero_mle}
\title{Zero handling function for MLE}
\usage{
zero_mle(sci_percentile, roll_mean, nn, n_zero, p0.center.mass)
}
\arguments{
\item{x}{Fill in}

\item{window}{Fill in}
}
\value{
An Lmoment fit.
}
\description{
Zero handling function for MLE
}
