% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpsites.R
\name{assess_dumpsites}
\alias{assess_dumpsites}
\title{Calculate dumpsite scores}
\usage{
assess_dumpsites(
  df_point,
  df_reach,
  start_date,
  end_date,
  reach_prefix_from_table,
  reach_prefix_from_layer,
  point_prefix_from_table,
  point_prefix_from_layer
)
}
\arguments{
\item{df_point}{Point data from Rapid Stream Assessment (RSA) dataset.}

\item{df_reach}{Reach data from Rapid Stream Assessment (RSA) dataset.}

\item{start_date}{Beginning of timeframe being evaluated (ex. first day of
five-year assessment period).}

\item{end_date}{End of timeframe being evaluated (ex. last day of
five-year assessment period).}

\item{reach_prefix_from_table}{Field name prefix sourced from RSA reach
attribute table. Should be consistent with the name of the stream reach
attribute table from the RSA geodatabase.}

\item{reach_prefix_from_layer}{Field name prefix sourced from RSA stream
reach layer. Should be consistent with the name of the stream reach
polyline layer from the RSA geodatabase.}

\item{point_prefix_from_table}{Field name prefix sourced from RSA point
attribute table. Should be consistent with the name of the point
assessment attribute table from the RSA geodatabase.}

\item{point_prefix_from_layer}{Field name prefix sourced from RSA point
layer. Should be consistent with the name of the intersected stream point
layer from the RSA geodatabase.}
}
\value{
List of 2 dataframes: summary and score.
}
\description{
Calculate dumpsite scores
}
\examples{
assess_dumpsites(df_point_example, df_reach_example, "2020-01-01", "2024-12-31","StreamReachAttributes", "StreamReaches","StreamPointAttributes", "StreamPoints_Intersect" )
}
