% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assign_score_from_maximum}
\alias{assign_score_from_maximum}
\title{Assign score based on lookup of maximum number. Used for dumpsites and EIA.}
\usage{
assign_score_from_maximum(value, lookup_table, lookup_field)
}
\arguments{
\item{value}{Number of dumpsites per mile for a given subwatershed.}

\item{lookup_table}{Lookup table of number of dumpsites per mile allowed per
score (1-10).}

\item{lookup_field}{Column name containing the maximum number of sites per
mile.}
}
\value{
Numerical score (1-10)
}
\description{
Assign score based on lookup of maximum number. Used for dumpsites and EIA.
}
\examples{
assign_score_from_maximum(0.5, dumpsite_score, "max_sites_per_mile")
}
