% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{calc_sci}
\alias{calc_sci}
\title{calc_sci}
\usage{
calc_sci(imgdf, mnidf, normdf = NULL, weights = 1)
}
\arguments{
\item{imgdf}{K x 2 data frame of K seeds and their respective brain
images to calculate the SCI on.}

\item{mnidf}{N x 4 data frame of K seeds and N MNI coordinates.}

\item{normdf}{N x 2 data frame to use for normalization.}

\item{weights}{N x 1 vector to use for weighting.}
}
\description{
This function calculates the striatal connectivity index (sci).
}
\examples{
calc_sci()
}
\keyword{connectivity,}
\keyword{correlation}
\keyword{functional}
\keyword{striatum,}
