% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{calc_sci}
\alias{calc_sci}
\title{calc_sci}
\usage{
calc_sci(img, vox, normdf = NULL, weights = 1)
}
\arguments{
\item{img}{brain image to calculate the sci on.}

\item{vox}{N x 3 matrix of voxel coordinates to use for regions of
interest.}

\item{normdf}{N x 2 data frame to use for normalization.}

\item{weights}{N x 1 vector to use for weighting.}
}
\description{
This function calculates the striatal connectivity index (sci).
}
\examples{
calc_sci()
}
\keyword{connectivity,}
\keyword{correlation}
\keyword{functional}
\keyword{striatum,}
