% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{get_val_nifti}
\alias{get_val_nifti}
\title{get_val_nifti}
\usage{
get_val_nifti(img, vox)
}
\arguments{
\item{img}{nifti object}

\item{vox}{voxel coordinates}
}
\description{
This function extracts a voxel value from a nifti object using the
fmri library in R
}
