% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCI-package.R
\docType{package}
\name{SCI-package}
\alias{SCI}
\alias{SCI-package}
\title{SCI: Standardized Climate Indices Such as SPI, SRI or SPEI}
\description{
Functions for generating Standardized Climate Indices (SCI). SCI is a transformation of (smoothed) climate (or environmental) time series that removes seasonality and forces the data to take values of the standard normal distribution. SCI was originally developed for precipitation. In this case it is known as the Standardized Precipitation Index (SPI).
}
\keyword{internal}
