% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCI-package.R
\docType{package}
\name{SCI-package}
\alias{SCI}
\alias{SCI-package}
\title{SCI: Stream Condition Index}
\description{
Calculates scores for hydrology, water quality, human health, and aquatic biology metrics to assess streams within the District of Columbia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LimnoTech/SCI}
  \item Report bugs at \url{https://github.com/LimnoTech/SCI/issues}
}

}
\author{
\strong{Maintainer}: Shanna Rucker \email{srucker@limno.com}

}
\keyword{internal}
